Squashed a Linux warning in wp::print(int64_t)

See merge request mmacklin/warp!59

+++ warp/native/builtin.h
@@ -532,7 +532,7 @@ inline CUDA_CALLABLE void print(int i)
 
 inline CUDA_CALLABLE void print(int64_t i)
 {
-    printf("%lld\n", i);
+    printf("%lld\n", (long long)i);
 }
 
 inline CUDA_CALLABLE void print(float f)
